@@ -1,3 +1,5 @@
+require "active_support/core_ext/integer/time"
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -15,52 +17,63 @@ Rails.application.configure do
   # Enable/disable caching. By default caching is disabled.
   if Rails.root.join('tmp/caching-dev.txt').exist?
     config.action_controller.perform_caching = true
-
-    config.cache_store = :dalli_store, { :pool_size => ENV.fetch("RAILS_MAX_THREADS") { 5 } }
-    config.public_file_server.headers = {
-      'Cache-Control' => "public, max-age=#{10.minutes.seconds.to_i}"
-    }
+    config.cache_store = :mem_cache_store, { pool_size: ENV.fetch("RAILS_MAX_THREADS", 5) }
+    config.public_file_server.headers = { 'Cache-Control' => "public, max-age=#{10.minutes.seconds.to_i}" }
   else
+    config.cache_store = :null_store
     config.action_controller.perform_caching = false
-
-    config.cache_store = :dalli_store, { :pool_size => ENV.fetch("RAILS_MAX_THREADS") { 5 } }
   end
 
   # Don't care if the mailer can't send.
-  config.action_mailer.raise_delivery_errors = false
+  # config.action_mailer.raise_delivery_errors = false
 
-  config.action_mailer.perform_caching = false
+  # config.action_mailer.perform_caching = false
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
+  # Raise exceptions for disallowed deprecations.
+  config.active_support.disallowed_deprecation = :raise
+
+  # Tell Active Support which deprecation messages to disallow.
+  config.active_support.disallowed_deprecation_warnings = []
+
   # Raise an error on page load if there are pending migrations.
   config.active_record.migration_error = :page_load
+
+  # Highlight code that triggered database queries in logs.
+  config.active_record.verbose_query_logs = true
 
   config.middleware.insert_before 0, Rack::Cors do
     allow do
       origins 'quill.org', %r{https://(.)*.quill.org}, /localhost:.*/, /127.0.0.1:.*/
 
       resource '*',
-        headers: :any,
-        methods: [:get, :post, :put, :patch, :delete, :options, :head],
-        credentials: true
+               headers: :any,
+               methods: %i[get post put patch delete options head],
+               credentials: true
     end
   end
 
   # Redis for caching
   if ENV['REDISCLOUD_URL']
     config.action_controller.perform_caching = true
-    config.cache_store = :redis_store, ENV["REDISCLOUD_URL"]
+    config.cache_store = :redis_store, ENV['REDISCLOUD_URL']
   else
     config.action_controller.perform_caching = false
   end
 
+  # Raises error for missing translations.
+  # config.i18n.raise_on_missing_translations = true
 
-  # Raises error for missing translations
-  # config.action_view.raise_on_missing_translations = true
+  # Annotate rendered view with file names.
+  # config.action_view.annotate_rendered_view_with_filenames = true
 
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
   config.file_watcher = ActiveSupport::EventedFileUpdateChecker
+
+  # Uncomment if you wish to allow Action Cable access from any origin.
+  # config.action_cable.disable_request_forgery_protection = true
 end
+
